fix(conversations): Keep transient work state out of SQL
Split the queryable conversation store from task execution state so SQL owns durable reporting records while pending mailbox payloads and leases stay in state. Rename the old metadata modules to conversation modules, remove the SQL inbound-message table, and update upgrade backfill/docs/tests around the shared Junior SQL database.

Refs GH-568
