+++ packages/docs/src/content/docs/cli/upgrade.md
@@ -58,8 +58,8 @@ Typical logs look like this:
 Running Junior upgrade migrations...
 Running migration migrate-legacy-conversation-work-redis-state...
 Finished migration migrate-legacy-conversation-work-redis-state: scanned=2 migrated=1 existing=0 missing=1
-Running migration backfill-conversation-metadata-sql...
-Finished migration backfill-conversation-metadata-sql: scanned=2 migrated=2 existing=0 missing=0
+Running migration backfill-conversations-sql...
+Finished migration backfill-conversations-sql: scanned=2 migrated=2 existing=0 missing=0
 Junior upgrade complete.
 ```
