+++ packages/junior/src/chat/app/production.ts
@@ -17,8 +17,8 @@ import type { VercelConversationWorkCallbackOptions } from "@/chat/task-executio
 import { resumeAwaitingSlackContinuation } from "@/chat/runtime/agent-continue-runner";
 import type { JuniorRuntimeServiceOverrides } from "@/chat/app/services";
 import { generateAssistantReply } from "@/chat/respond";
-import { getConfiguredConversationMetadataStore } from "@/chat/metadata/configured-store";
-import type { ConversationMetadataStore } from "@/chat/metadata/store";
+import { getConfiguredConversationStore } from "@/chat/conversations/configured";
+import type { ConversationStore } from "@/chat/conversations/store";
 
 let productionSlackAdapter: SlackAdapter | undefined;
 let productionSlackRuntime: ReturnType<typeof createSlackRuntime> | undefined;
@@ -58,32 +58,32 @@ export function getProductionSlackRuntime(): ReturnType<
   return productionSlackRuntime;
 }
 
-/** Return the production conversation metadata store for current config. */
-export function getProductionConversationMetadataStore(): ConversationMetadataStore {
-  return getConfiguredConversationMetadataStore();
+/** Return the production conversation store for current config. */
+export function getProductionConversationStore(): ConversationStore {
+  return getConfiguredConversationStore();
 }
 
 /** Create production-backed services for Slack webhook ingress. */
 export function createProductionSlackWebhookServices(options?: {
   services?: JuniorRuntimeServiceOverrides;
 }): SlackWebhookServices {
-  const metadataStore =
-    options?.services?.replyExecutor?.metadataStore ??
-    getProductionConversationMetadataStore();
+  const conversationStore =
+    options?.services?.replyExecutor?.conversationStore ??
+    getProductionConversationStore();
   const runtime = createSlackRuntime({
     getSlackAdapter: getProductionSlackAdapter,
     services: {
       ...options?.services,
       replyExecutor: {
         ...options?.services?.replyExecutor,
-        metadataStore,
+        conversationStore,
       },
     },
   });
   return {
     getSlackAdapter: getProductionSlackAdapter,
     getUserTokenStore: createUserTokenStore,
-    metadataStore,
+    conversationStore,
     queue: getVercelConversationWorkQueue(),
     runtime,
   };
@@ -91,11 +91,11 @@ export function createProductionSlackWebhookServices(options?: {
 
 /** Return production services for Slack webhook ingress. */
 export function getProductionSlackWebhookServices(): SlackWebhookServices {
-  const metadataStore = getProductionConversationMetadataStore();
+  const conversationStore = getProductionConversationStore();
   return {
     getSlackAdapter: getProductionSlackAdapter,
     getUserTokenStore: createUserTokenStore,
-    metadataStore,
+    conversationStore,
     queue: getVercelConversationWorkQueue(),
     runtime: getProductionSlackRuntime(),
   };
@@ -105,25 +105,25 @@ export function getProductionSlackWebhookServices(): SlackWebhookServices {
 export function createProductionConversationWorkOptions(options?: {
   services?: JuniorRuntimeServiceOverrides;
 }): VercelConversationWorkCallbackOptions {
-  const metadataStore =
-    options?.services?.replyExecutor?.metadataStore ??
-    getProductionConversationMetadataStore();
+  const conversationStore =
+    options?.services?.replyExecutor?.conversationStore ??
+    getProductionConversationStore();
   const runtime = createSlackRuntime({
     getSlackAdapter: getProductionSlackAdapter,
     services: {
       ...options?.services,
       replyExecutor: {
         ...options?.services?.replyExecutor,
-        metadataStore,
+        conversationStore,
       },
     },
   });
   return {
-    metadataStore,
+    conversationStore,
     queue: getVercelConversationWorkQueue(),
     run: createSlackConversationWorker({
       getSlackAdapter: getProductionSlackAdapter,
-      metadataStore,
+      conversationStore,
       resumeAwaitingContinuation: async (conversationId) =>
         await resumeAwaitingSlackContinuation(conversationId, {
           generateReply: withSandboxTracePropagation(
