+++ packages/junior/src/chat/conversations/configured.ts
@@ -1,20 +1,13 @@
-/**
- * Process-configured conversation metadata store boundary.
- *
- * Production uses SQL when a database URL is configured; local/default runs use
- * the state-backed adapter. Cache SQL stores by URL so Neon pools are reused
- * without freezing test configuration changes.
- */
 import { getChatConfig } from "@/chat/config";
 import { createNeonJuniorSqlExecutor } from "@/chat/sql/neon";
-import { createStateConversationMetadataStore } from "./state-store";
+import { createStateConversationStore } from "./state";
 import { createSqlStore } from "./sql/store";
-import type { ConversationMetadataStore } from "./store";
+import type { ConversationStore } from "./store";
 
 let configuredStore:
   | {
       databaseUrl: string;
-      store: ConversationMetadataStore;
+      store: ConversationStore;
     }
   | undefined;
 
@@ -23,11 +16,11 @@ export function hasConfiguredJuniorDatabase(): boolean {
   return Boolean(getChatConfig().sql.databaseUrl);
 }
 
-/** Return the process-configured conversation metadata store. */
-export function getConfiguredConversationMetadataStore(): ConversationMetadataStore {
+/** Return the process-configured conversation feed store. */
+export function getConfiguredConversationStore(): ConversationStore {
   const databaseUrl = getChatConfig().sql.databaseUrl;
   if (!databaseUrl) {
-    return createStateConversationMetadataStore();
+    return createStateConversationStore();
   }
   if (configuredStore?.databaseUrl !== databaseUrl) {
     configuredStore = {
