+++ packages/junior/src/chat/conversations/sql/schema.ts
@@ -1,13 +1,9 @@
-import {
-  juniorConversationInboundMessages,
-  juniorConversations,
-} from "./schema/conversations";
+import { juniorConversations } from "./schema/conversations";
 import { juniorDestinations } from "./schema/destinations";
 import { juniorIdentities } from "./schema/identities";
 import { juniorSchemaMigrations } from "./schema/migrations";
 
 export {
-  juniorConversationInboundMessages,
   juniorConversations,
   juniorDestinations,
   juniorIdentities,
@@ -15,7 +11,6 @@ export {
 };
 
 export const schema = {
-  juniorConversationInboundMessages,
   juniorConversations,
   juniorDestinations,
   juniorIdentities,
