+++ packages/junior/src/chat/conversations/sql/schema/migrations.ts
