+++ packages/junior/src/chat/conversations/sql/store.ts
@@ -1,0 +1,724 @@
+import { randomUUID } from "node:crypto";
+import type { Destination } from "@sentry/junior-plugin-api";
+import { asc, desc, eq, sql } from "drizzle-orm";
+import { parseDestination, sameDestination } from "@/chat/destination";
+import {
+  parseStoredSlackRequester,
+  type StoredSlackRequester,
+} from "@/chat/requester";
+import {
+  type Conversation,
+  type ConversationExecution,
+  type ExecutionStatus,
+  type Lease,
+  type Source,
+} from "@/chat/task-execution/state";
+import { migrateSchema } from "./migrations";
+import type {
+  JuniorSqlDatabase,
+  JuniorSqlMigrationExecutor,
+} from "@/chat/sql/db";
+import type { ConversationStore } from "../store";
+import {
+  juniorConversations,
+  juniorDestinations,
+  juniorIdentities,
+} from "./schema";
+import type {
+  JuniorDestinationKind,
+  JuniorDestinationVisibility,
+} from "./schema/destinations";
+import type { JuniorIdentityKind } from "./schema/identities";
+
+type ConversationRow = typeof juniorConversations.$inferSelect;
+
+interface IdentityUpsert {
+  email?: string;
+  displayName?: string;
+  handle?: string;
+  kind: JuniorIdentityKind;
+  metadata?: Record<string, unknown>;
+  provider: string;
+  providerSubjectId: string;
+  providerTenantId?: string;
+}
+
+interface DestinationUpsert {
+  displayName?: string;
+  kind: JuniorDestinationKind;
+  metadata?: Record<string, unknown>;
+  provider: string;
+  providerDestinationId: string;
+  providerTenantId?: string;
+  visibility: JuniorDestinationVisibility;
+}
+
+const CONVERSATION_MUTATION_LOCK_PREFIX = "junior_conversation";
+
+function now(): number {
+  return Date.now();
+}
+
+function dateFromMs(ms: number): Date {
+  return new Date(ms);
+}
+
+function tenantId(value: string | undefined): string {
+  return value ?? "";
+}
+
+function msFromDate(
+  value: Date | string | null | undefined,
+): number | undefined {
+  if (value === null || value === undefined) {
+    return undefined;
+  }
+  const date = value instanceof Date ? value : new Date(value);
+  return date.getTime();
+}
+
+function requiredMsFromDate(value: Date | string): number {
+  const ms = msFromDate(value);
+  if (typeof ms !== "number" || Number.isNaN(ms)) {
+    throw new Error("Conversation record timestamp is invalid");
+  }
+  return ms;
+}
+
+function sourceFromValue(value: unknown): Source | undefined {
+  if (
+    value === "api" ||
+    value === "internal" ||
+    value === "local" ||
+    value === "plugin" ||
+    value === "scheduler" ||
+    value === "slack"
+  ) {
+    return value;
+  }
+  return undefined;
+}
+
+function identityFromRequester(
+  requester: StoredSlackRequester | undefined,
+): IdentityUpsert | undefined {
+  if (!requester?.slackUserId) {
+    return undefined;
+  }
+  return {
+    kind: "user",
+    provider: "slack",
+    providerTenantId: requester.teamId,
+    providerSubjectId: requester.slackUserId,
+    ...(requester.fullName ? { displayName: requester.fullName } : {}),
+    ...(requester.slackUserName ? { handle: requester.slackUserName } : {}),
+    ...(requester.email ? { email: requester.email } : {}),
+    metadata: { platform: "slack" },
+  };
+}
+
+function systemIdentityFromSource(
+  source: Source | undefined,
+): IdentityUpsert | undefined {
+  if (source === "scheduler") {
+    return {
+      kind: "system",
+      provider: "junior",
+      providerSubjectId: "scheduler",
+      displayName: "Junior Scheduler",
+    };
+  }
+  if (source === "local") {
+    return {
+      kind: "system",
+      provider: "junior",
+      providerSubjectId: "local-cli",
+      displayName: "Local CLI",
+    };
+  }
+  return undefined;
+}
+
+function actorIdentityForConversation(
+  conversation: Conversation,
+): IdentityUpsert | undefined {
+  return (
+    identityFromRequester(conversation.requester) ??
+    systemIdentityFromSource(conversation.source)
+  );
+}
+
+function originTypeFromSource(source: Source | undefined): string | undefined {
+  return source;
+}
+
+function localWorkspaceFromConversationId(
+  conversationId: string,
+): string | undefined {
+  const match = /^local:([^:]+):/.exec(conversationId);
+  return match?.[1];
+}
+
+function destinationUpsertFromDestination(args: {
+  channelName?: string;
+  conversationId?: string;
+  destination: Destination | undefined;
+}): DestinationUpsert | undefined {
+  const { destination } = args;
+  if (!destination) {
+    return undefined;
+  }
+  if (destination.platform === "slack") {
+    const channelId = destination.channelId;
+    const channelKind = channelId.startsWith("D")
+      ? "dm"
+      : channelId.startsWith("G")
+        ? "group"
+        : "channel";
+    const visibility = channelId.startsWith("D")
+      ? "direct"
+      : channelId.startsWith("G")
+        ? "private"
+        : "public";
+    return {
+      kind: channelKind,
+      provider: "slack",
+      providerTenantId: destination.teamId,
+      providerDestinationId: channelId,
+      visibility,
+      ...(args.channelName ? { displayName: args.channelName } : {}),
+      metadata: { platform: "slack" },
+    };
+  }
+  return {
+    kind: "local_conversation",
+    provider: "local",
+    providerTenantId:
+      localWorkspaceFromConversationId(destination.conversationId) ??
+      localWorkspaceFromConversationId(args.conversationId ?? ""),
+    providerDestinationId: destination.conversationId,
+    visibility: "direct",
+    metadata: { platform: "local" },
+  };
+}
+
+function executionStatusFromValue(value: unknown): ExecutionStatus {
+  if (
+    value === "awaiting_resume" ||
+    value === "idle" ||
+    value === "pending" ||
+    value === "running"
+  ) {
+    return value;
+  }
+  throw new Error("Conversation record execution status is invalid");
+}
+
+function leaseFromRow(row: ConversationRow): Lease | undefined {
+  if (!row.leaseToken) {
+    return undefined;
+  }
+  const acquiredAtMs = msFromDate(row.leaseAcquiredAt);
+  const lastCheckInAtMs = msFromDate(row.leaseLastCheckInAt);
+  const expiresAtMs = msFromDate(row.leaseExpiresAt);
+  if (
+    typeof acquiredAtMs !== "number" ||
+    typeof lastCheckInAtMs !== "number" ||
+    typeof expiresAtMs !== "number"
+  ) {
+    return undefined;
+  }
+  return {
+    token: row.leaseToken,
+    acquiredAtMs,
+    lastCheckInAtMs,
+    expiresAtMs,
+  };
+}
+
+function conversationFromRow(row: ConversationRow): Conversation {
+  if (row.schemaVersion !== 1) {
+    throw new Error("Conversation record schema version is invalid");
+  }
+  const destination =
+    row.destination === undefined || row.destination === null
+      ? undefined
+      : parseDestination(row.destination);
+  const requester = parseStoredSlackRequester(row.requester);
+  const source =
+    row.source === undefined || row.source === null
+      ? undefined
+      : sourceFromValue(row.source);
+  if (row.source !== undefined && row.source !== null && !source) {
+    throw new Error("Conversation record source is invalid");
+  }
+  const lease = leaseFromRow(row);
+  const execution: ConversationExecution = {
+    status: executionStatusFromValue(row.executionStatus),
+    inboundMessageIds: [],
+    pendingCount: 0,
+    pendingMessages: [],
+    lastCheckpointAtMs: msFromDate(row.lastCheckpointAt),
+    lastEnqueuedAtMs: msFromDate(row.lastEnqueuedAt),
+    ...(lease ? { lease } : {}),
+    ...(row.runId ? { runId: row.runId } : {}),
+    updatedAtMs:
+      msFromDate(row.executionUpdatedAt) ?? requiredMsFromDate(row.updatedAt),
+  };
+
+  return {
+    schemaVersion: 1,
+    conversationId: row.conversationId,
+    createdAtMs: requiredMsFromDate(row.createdAt),
+    lastActivityAtMs: requiredMsFromDate(row.lastActivityAt),
+    updatedAtMs: requiredMsFromDate(row.updatedAt),
+    execution,
+    ...(destination ? { destination } : {}),
+    ...(requester ? { requester } : {}),
+    ...(row.channelName ? { channelName: row.channelName } : {}),
+    ...(source ? { source } : {}),
+    ...(row.title ? { title: row.title } : {}),
+  };
+}
+
+function emptyConversation(args: {
+  conversationId: string;
+  destination?: Destination;
+  nowMs: number;
+  source?: Source;
+}): Conversation {
+  return {
+    schemaVersion: 1,
+    conversationId: args.conversationId,
+    createdAtMs: args.nowMs,
+    lastActivityAtMs: args.nowMs,
+    updatedAtMs: args.nowMs,
+    ...(args.destination ? { destination: args.destination } : {}),
+    ...(args.source ? { source: args.source } : {}),
+    execution: {
+      status: "idle",
+      inboundMessageIds: [],
+      pendingCount: 0,
+      pendingMessages: [],
+      updatedAtMs: args.nowMs,
+    },
+  };
+}
+
+function assertSameConversationDestination(args: {
+  conversationId: string;
+  current: Destination | undefined;
+  next: Destination;
+}): void {
+  if (!args.current || sameDestination(args.current, args.next)) {
+    return;
+  }
+  throw new Error(
+    `Conversation destination changed for ${args.conversationId}`,
+  );
+}
+
+export class SqlStore implements ConversationStore {
+  private schemaReady: Promise<void> | undefined;
+
+  constructor(
+    private readonly executor: JuniorSqlDatabase,
+    private readonly migrationExecutor: JuniorSqlMigrationExecutor,
+  ) {}
+
+  /** Apply SQL schema migrations before runtime uses this store. */
+  async migrate(): Promise<void> {
+    if (!this.schemaReady) {
+      this.schemaReady = migrateSchema(this.migrationExecutor);
+    }
+    const schemaReady = this.schemaReady;
+    try {
+      await schemaReady;
+    } catch (error) {
+      if (this.schemaReady === schemaReady) {
+        this.schemaReady = undefined;
+      }
+      throw error;
+    }
+  }
+
+  async getConversation(args: {
+    conversationId: string;
+  }): Promise<Conversation | undefined> {
+    const row = await this.readConversationRow(args.conversationId);
+    if (!row) {
+      return undefined;
+    }
+    return conversationFromRow(row);
+  }
+
+  async recordConversationActivity(args: {
+    activityAtMs?: number;
+    channelName?: string;
+    conversationId: string;
+    destination?: Destination;
+    nowMs?: number;
+    requester?: StoredSlackRequester;
+    source?: Source;
+    title?: string;
+  }): Promise<void> {
+    const nowMs = args.nowMs ?? now();
+    const activityAtMs = args.activityAtMs ?? nowMs;
+    await this.withConversationMutation(args.conversationId, async () => {
+      const existing = await this.getConversation({
+        conversationId: args.conversationId,
+      });
+      if (existing && args.destination) {
+        assertSameConversationDestination({
+          conversationId: args.conversationId,
+          current: existing.destination,
+          next: args.destination,
+        });
+      }
+      const current =
+        existing ??
+        emptyConversation({
+          conversationId: args.conversationId,
+          destination: args.destination,
+          nowMs,
+          source: args.source,
+        });
+      await this.upsertConversation({
+        conversation: {
+          ...current,
+          destination: current.destination ?? args.destination,
+          source: current.source ?? args.source,
+          channelName: current.channelName ?? args.channelName,
+          requester: current.requester ?? args.requester,
+          title: current.title ?? args.title,
+          lastActivityAtMs: Math.max(current.lastActivityAtMs, activityAtMs),
+          updatedAtMs: nowMs,
+          execution: {
+            ...current.execution,
+            updatedAtMs: current.execution.updatedAtMs ?? nowMs,
+          },
+        },
+      });
+    });
+  }
+
+  async recordConversationStatus(args: {
+    channelName?: string;
+    conversationId: string;
+    createdAtMs: number;
+    destination?: Destination;
+    execution: Pick<
+      ConversationExecution,
+      | "lastCheckpointAtMs"
+      | "lastEnqueuedAtMs"
+      | "lease"
+      | "runId"
+      | "status"
+      | "updatedAtMs"
+    >;
+    lastActivityAtMs: number;
+    requester?: StoredSlackRequester;
+    source?: Source;
+    title?: string;
+    updatedAtMs: number;
+  }): Promise<void> {
+    await this.withConversationMutation(args.conversationId, async () => {
+      await this.upsertConversation({
+        conversation: {
+          schemaVersion: 1,
+          conversationId: args.conversationId,
+          createdAtMs: args.createdAtMs,
+          lastActivityAtMs: args.lastActivityAtMs,
+          updatedAtMs: args.updatedAtMs,
+          ...(args.channelName ? { channelName: args.channelName } : {}),
+          ...(args.destination ? { destination: args.destination } : {}),
+          ...(args.requester ? { requester: args.requester } : {}),
+          ...(args.source ? { source: args.source } : {}),
+          ...(args.title ? { title: args.title } : {}),
+          execution: {
+            inboundMessageIds: [],
+            pendingCount: 0,
+            pendingMessages: [],
+            ...args.execution,
+          },
+        },
+      });
+    });
+  }
+
+  /** Copy one conversation record into SQL during backfill. */
+  async backfillConversation(conversation: Conversation): Promise<void> {
+    await this.withConversationMutation(
+      conversation.conversationId,
+      async () => {
+        const existing = await this.getConversation({
+          conversationId: conversation.conversationId,
+        });
+        const sourceExecutionAtMs =
+          conversation.execution.updatedAtMs ?? conversation.updatedAtMs;
+        const existingExecutionAtMs =
+          existing === undefined
+            ? undefined
+            : (existing.execution.updatedAtMs ?? existing.updatedAtMs);
+        const refreshExecutionFromSource =
+          existingExecutionAtMs === undefined ||
+          sourceExecutionAtMs >= existingExecutionAtMs;
+        const mergedConversation = existing
+          ? {
+              ...conversation,
+              channelName: existing.channelName ?? conversation.channelName,
+              createdAtMs: Math.min(
+                existing.createdAtMs,
+                conversation.createdAtMs,
+              ),
+              destination: existing.destination ?? conversation.destination,
+              lastActivityAtMs: Math.max(
+                existing.lastActivityAtMs,
+                conversation.lastActivityAtMs,
+              ),
+              requester: existing.requester ?? conversation.requester,
+              source: existing.source ?? conversation.source,
+              title: existing.title ?? conversation.title,
+              updatedAtMs: Math.max(
+                existing.updatedAtMs,
+                conversation.updatedAtMs,
+              ),
+              execution: refreshExecutionFromSource
+                ? conversation.execution
+                : existing.execution,
+            }
+          : conversation;
+        await this.upsertConversation({ conversation: mergedConversation });
+      },
+    );
+  }
+
+  async listConversationsByActivity(
+    args: {
+      limit?: number;
+      offset?: number;
+    } = {},
+  ): Promise<Conversation[]> {
+    const rows = await this.executor
+      .db()
+      .select()
+      .from(juniorConversations)
+      .orderBy(
+        desc(juniorConversations.lastActivityAt),
+        asc(juniorConversations.conversationId),
+      )
+      .limit(Math.max(0, args.limit ?? 10_000))
+      .offset(Math.max(0, args.offset ?? 0));
+    const conversations: Conversation[] = [];
+    for (const row of rows) {
+      conversations.push(conversationFromRow(row));
+    }
+    return conversations;
+  }
+
+  /** Serialize all durable mutations for one conversation inside a SQL transaction. */
+  private async withConversationMutation<T>(
+    conversationId: string,
+    callback: () => Promise<T>,
+  ): Promise<T> {
+    return await this.executor.withLock(
+      `${CONVERSATION_MUTATION_LOCK_PREFIX}:${conversationId}`,
+      async () => await this.executor.transaction(callback),
+    );
+  }
+
+  private async readConversationRow(
+    conversationId: string,
+  ): Promise<ConversationRow | undefined> {
+    const rows = await this.executor
+      .db()
+      .select()
+      .from(juniorConversations)
+      .where(eq(juniorConversations.conversationId, conversationId));
+    return rows[0];
+  }
+
+  /** Upsert the conversation row while preserving previously discovered nullable metadata fields. */
+  private async upsertConversation(args: {
+    conversation: Conversation;
+  }): Promise<void> {
+    const { conversation } = args;
+    const lease = conversation.execution.lease;
+    const destinationId = await this.upsertDestination(
+      destinationUpsertFromDestination({
+        channelName: conversation.channelName,
+        conversationId: conversation.conversationId,
+        destination: conversation.destination,
+      }),
+      conversation.updatedAtMs,
+    );
+    const requesterIdentityId = await this.upsertIdentity(
+      identityFromRequester(conversation.requester),
+      conversation.updatedAtMs,
+    );
+    const actorIdentityId = await this.upsertIdentity(
+      actorIdentityForConversation(conversation),
+      conversation.updatedAtMs,
+    );
+    await this.executor
+      .db()
+      .insert(juniorConversations)
+      .values({
+        conversationId: conversation.conversationId,
+        schemaVersion: 1,
+        source: conversation.source ?? null,
+        originType: originTypeFromSource(conversation.source) ?? null,
+        originId: null,
+        originRunId: null,
+        destinationId: destinationId ?? null,
+        destination: conversation.destination ?? null,
+        actorIdentityId: actorIdentityId ?? null,
+        requesterIdentityId: requesterIdentityId ?? null,
+        creatorIdentityId: null,
+        credentialSubjectIdentityId: null,
+        requester: conversation.requester ?? null,
+        channelName: conversation.channelName ?? null,
+        title: conversation.title ?? null,
+        createdAt: dateFromMs(conversation.createdAtMs),
+        lastActivityAt: dateFromMs(conversation.lastActivityAtMs),
+        updatedAt: dateFromMs(conversation.updatedAtMs),
+        executionUpdatedAt:
+          conversation.execution.updatedAtMs === undefined
+            ? null
+            : dateFromMs(conversation.execution.updatedAtMs),
+        executionStatus: conversation.execution.status,
+        runId: conversation.execution.runId ?? null,
+        lastCheckpointAt:
+          conversation.execution.lastCheckpointAtMs === undefined
+            ? null
+            : dateFromMs(conversation.execution.lastCheckpointAtMs),
+        lastEnqueuedAt:
+          conversation.execution.lastEnqueuedAtMs === undefined
+            ? null
+            : dateFromMs(conversation.execution.lastEnqueuedAtMs),
+        leaseToken: lease?.token ?? null,
+        leaseAcquiredAt: lease ? dateFromMs(lease.acquiredAtMs) : null,
+        leaseLastCheckInAt: lease ? dateFromMs(lease.lastCheckInAtMs) : null,
+        leaseExpiresAt: lease ? dateFromMs(lease.expiresAtMs) : null,
+      })
+      .onConflictDoUpdate({
+        target: juniorConversations.conversationId,
+        set: {
+          source: sql`coalesce(excluded.source, ${juniorConversations.source})`,
+          originType: sql`coalesce(excluded.origin_type, ${juniorConversations.originType})`,
+          originId: sql`coalesce(excluded.origin_id, ${juniorConversations.originId})`,
+          originRunId: sql`coalesce(excluded.origin_run_id, ${juniorConversations.originRunId})`,
+          destinationId: sql`coalesce(excluded.destination_id, ${juniorConversations.destinationId})`,
+          destination: sql`coalesce(excluded.destination_json, ${juniorConversations.destination})`,
+          actorIdentityId: sql`coalesce(excluded.actor_identity_id, ${juniorConversations.actorIdentityId})`,
+          requesterIdentityId: sql`coalesce(excluded.requester_identity_id, ${juniorConversations.requesterIdentityId})`,
+          creatorIdentityId: sql`coalesce(excluded.creator_identity_id, ${juniorConversations.creatorIdentityId})`,
+          credentialSubjectIdentityId: sql`coalesce(excluded.credential_subject_identity_id, ${juniorConversations.credentialSubjectIdentityId})`,
+          requester: sql`coalesce(excluded.requester_json, ${juniorConversations.requester})`,
+          channelName: sql`coalesce(excluded.channel_name, ${juniorConversations.channelName})`,
+          title: sql`coalesce(excluded.title, ${juniorConversations.title})`,
+          lastActivityAt: sql`greatest(${juniorConversations.lastActivityAt}, excluded.last_activity_at)`,
+          updatedAt: sql`excluded.updated_at`,
+          executionUpdatedAt: sql`excluded.execution_updated_at`,
+          executionStatus: sql`excluded.execution_status`,
+          runId: sql`excluded.run_id`,
+          lastCheckpointAt: sql`excluded.last_checkpoint_at`,
+          lastEnqueuedAt: sql`excluded.last_enqueued_at`,
+          leaseToken: sql`excluded.lease_token`,
+          leaseAcquiredAt: sql`excluded.lease_acquired_at`,
+          leaseLastCheckInAt: sql`excluded.lease_last_check_in_at`,
+          leaseExpiresAt: sql`excluded.lease_expires_at`,
+        },
+      });
+  }
+
+  private async upsertIdentity(
+    identity: IdentityUpsert | undefined,
+    nowMs: number,
+  ): Promise<string | undefined> {
+    if (!identity) {
+      return undefined;
+    }
+    const rows = await this.executor
+      .db()
+      .insert(juniorIdentities)
+      .values({
+        id: randomUUID(),
+        kind: identity.kind,
+        provider: identity.provider,
+        providerTenantId: tenantId(identity.providerTenantId),
+        providerSubjectId: identity.providerSubjectId,
+        displayName: identity.displayName ?? null,
+        handle: identity.handle ?? null,
+        email: identity.email ?? null,
+        avatarUrl: null,
+        metadata: identity.metadata ?? null,
+        createdAt: dateFromMs(nowMs),
+        updatedAt: dateFromMs(nowMs),
+      })
+      .onConflictDoUpdate({
+        target: [
+          juniorIdentities.provider,
+          juniorIdentities.providerTenantId,
+          juniorIdentities.providerSubjectId,
+        ],
+        set: {
+          kind: sql`excluded.kind`,
+          displayName: sql`coalesce(excluded.display_name, ${juniorIdentities.displayName})`,
+          handle: sql`coalesce(excluded.handle, ${juniorIdentities.handle})`,
+          email: sql`coalesce(excluded.email, ${juniorIdentities.email})`,
+          avatarUrl: sql`coalesce(excluded.avatar_url, ${juniorIdentities.avatarUrl})`,
+          metadata: sql`coalesce(excluded.metadata_json, ${juniorIdentities.metadata})`,
+          updatedAt: sql`excluded.updated_at`,
+        },
+      })
+      .returning({ id: juniorIdentities.id });
+    return rows[0]?.id;
+  }
+
+  private async upsertDestination(
+    destination: DestinationUpsert | undefined,
+    nowMs: number,
+  ): Promise<string | undefined> {
+    if (!destination) {
+      return undefined;
+    }
+    const rows = await this.executor
+      .db()
+      .insert(juniorDestinations)
+      .values({
+        id: randomUUID(),
+        provider: destination.provider,
+        providerTenantId: tenantId(destination.providerTenantId),
+        providerDestinationId: destination.providerDestinationId,
+        kind: destination.kind,
+        parentDestinationId: null,
+        displayName: destination.displayName ?? null,
+        visibility: destination.visibility,
+        metadata: destination.metadata ?? null,
+        createdAt: dateFromMs(nowMs),
+        updatedAt: dateFromMs(nowMs),
+      })
+      .onConflictDoUpdate({
+        target: [
+          juniorDestinations.provider,
+          juniorDestinations.providerTenantId,
+          juniorDestinations.providerDestinationId,
+        ],
+        set: {
+          kind: sql`excluded.kind`,
+          displayName: sql`coalesce(excluded.display_name, ${juniorDestinations.displayName})`,
+          visibility: sql`excluded.visibility`,
+          metadata: sql`coalesce(excluded.metadata_json, ${juniorDestinations.metadata})`,
+          updatedAt: sql`excluded.updated_at`,
+        },
+      })
+      .returning({ id: juniorDestinations.id });
+    return rows[0]?.id;
+  }
+}
+
+/** Create a SQL-backed conversation store. */
+export function createSqlStore(executor: JuniorSqlMigrationExecutor): SqlStore {
+  return new SqlStore(executor, executor);
+}
