+++ packages/junior/src/chat/ingress/slack-webhook.ts
@@ -7,7 +7,7 @@ import {
   type StateAdapter,
 } from "chat";
 import type { SlackTurnRuntime } from "@/chat/runtime/slack-runtime";
-import type { ConversationMetadataStore } from "@/chat/metadata/store";
+import type { ConversationStore } from "@/chat/conversations/store";
 import type { ConversationWorkQueue } from "@/chat/task-execution/queue";
 import { appendAndEnqueueInboundMessage } from "@/chat/task-execution/store";
 import {
@@ -95,7 +95,7 @@ export interface SlackWebhookServices {
   getUserTokenStore?: () => UserTokenStore;
   getSlackAdapter: () => SlackAdapter;
   queue: ConversationWorkQueue;
-  metadataStore?: ConversationMetadataStore;
+  conversationStore?: ConversationStore;
   runtime: Pick<
     SlackTurnRuntime<unknown>,
     | "handleAssistantContextChanged"
@@ -190,7 +190,7 @@ async function persistSlackMessage(args: {
   adapter: SlackAdapter;
   installation: SlackInstallationContext;
   message: Message;
-  metadataStore?: ConversationMetadataStore;
+  conversationStore?: ConversationStore;
   queue: ConversationWorkQueue;
   receivedAtMs: number;
   route: SlackConversationRoute;
@@ -207,7 +207,7 @@ async function persistSlackMessage(args: {
   });
   await appendAndEnqueueInboundMessage({
     message: inbound,
-    metadataStore: args.metadataStore,
+    conversationStore: args.conversationStore,
     queue: args.queue,
     state: args.state,
   });
@@ -218,7 +218,7 @@ async function routeParsedMessage(args: {
   event: SlackMessageEvent;
   installation: SlackInstallationContext;
   message: Message;
-  metadataStore?: ConversationMetadataStore;
+  conversationStore?: ConversationStore;
   queue: ConversationWorkQueue;
   receivedAtMs: number;
   state: StateAdapter;
@@ -249,7 +249,7 @@ async function routeParsedMessage(args: {
     adapter: args.adapter,
     installation: args.installation,
     message: args.message,
-    metadataStore: args.metadataStore,
+    conversationStore: args.conversationStore,
     queue: args.queue,
     receivedAtMs: args.receivedAtMs,
     route,
@@ -262,7 +262,7 @@ async function handleMessageChanged(args: {
   body: unknown;
   installation: SlackInstallationContext;
   queue: ConversationWorkQueue;
-  metadataStore?: ConversationMetadataStore;
+  conversationStore?: ConversationStore;
   receivedAtMs: number;
   state: StateAdapter;
 }): Promise<boolean> {
@@ -287,7 +287,7 @@ async function handleMessageChanged(args: {
     adapter: args.adapter,
     installation: args.installation,
     message: result.message,
-    metadataStore: args.metadataStore,
+    conversationStore: args.conversationStore,
     queue: args.queue,
     receivedAtMs: args.receivedAtMs,
     route: "mention",
@@ -336,7 +336,7 @@ async function handleSlackEvent(args: {
             adapter,
             body: args.body,
             installation,
-            metadataStore: args.services.metadataStore,
+            conversationStore: args.services.conversationStore,
             queue: args.services.queue,
             receivedAtMs,
             state,
@@ -412,7 +412,7 @@ async function handleSlackEvent(args: {
             event,
             installation,
             message,
-            metadataStore: args.services.metadataStore,
+            conversationStore: args.services.conversationStore,
             queue: args.services.queue,
             receivedAtMs,
             state,
