+++ packages/junior/src/chat/respond.ts
@@ -150,7 +150,7 @@ import {
   toGenAiMessageMetadata,
   toGenAiMessagesTraceAttributes,
 } from "@/chat/conversation-privacy";
-import type { ConversationMetadataStore } from "@/chat/metadata/store";
+import type { ConversationStore } from "@/chat/conversations/store";
 
 // Re-export types for backward compatibility with existing consumers.
 export type { AssistantReply, AgentTurnDiagnostics };
@@ -198,7 +198,7 @@ function waitForAbortSettlement(
 export interface ReplyRequestContext {
   skillDirs?: string[];
   credentialContext?: CredentialContext;
-  metadataStore?: ConversationMetadataStore;
+  conversationStore?: ConversationStore;
   requester?: Requester;
   slackConversation?: SlackConversationContext;
   destination: Destination;
@@ -1286,7 +1286,7 @@ export async function generateAssistantReply(
         messages,
         loadedSkillNames: loadedSkillNamesForResume,
         logContext: sessionRecordLogContext,
-        metadataStore: context.metadataStore,
+        conversationStore: context.conversationStore,
         requester,
         ...(surface ? { surface } : {}),
         ...(turnStartMessageIndex !== undefined
@@ -1663,7 +1663,7 @@ export async function generateAssistantReply(
         allMessages: agent.state.messages,
         loadedSkillNames: loadedSkillNamesForResume,
         logContext: sessionRecordLogContext,
-        metadataStore: context.metadataStore,
+        conversationStore: context.conversationStore,
         requester,
         ...(surface ? { surface } : {}),
         ...(turnStartMessageIndex !== undefined
@@ -1715,7 +1715,7 @@ export async function generateAssistantReply(
         errorMessage: error.message,
         loadedSkillNames: loadedSkillNamesForResume,
         logContext: sessionRecordLogContext,
-        metadataStore: context.metadataStore,
+        conversationStore: context.conversationStore,
         requester,
         ...(surface ? { surface } : {}),
       });
@@ -1744,7 +1744,7 @@ export async function generateAssistantReply(
         errorMessage: error instanceof Error ? error.message : String(error),
         loadedSkillNames: loadedSkillNamesForResume,
         logContext: sessionRecordLogContext,
-        metadataStore: context.metadataStore,
+        conversationStore: context.conversationStore,
         requester,
         ...(surface ? { surface } : {}),
       });
@@ -1796,7 +1796,7 @@ export async function generateAssistantReply(
         errorMessage: error.message,
         loadedSkillNames: loadedSkillNamesForResume,
         logContext: sessionRecordLogContext,
-        metadataStore: context.metadataStore,
+        conversationStore: context.conversationStore,
         requester,
         ...(surface ? { surface } : {}),
       });
