+++ packages/junior/src/chat/task-execution/slack-work.ts
@@ -16,7 +16,7 @@ import {
 import { normalizeIncomingSlackThreadId } from "@/chat/ingress/message-router";
 import { rehydrateAttachmentFetchers } from "@/chat/slack/attachment-fetchers";
 import { getStateAdapter } from "@/chat/state/adapter";
-import type { ConversationMetadataStore } from "@/chat/metadata/store";
+import type { ConversationStore } from "@/chat/conversations/store";
 import type { AgentInput, InboundMessage } from "@/chat/task-execution/store";
 import {
   getConversationWorkState,
@@ -56,7 +56,7 @@ export interface CreateSlackConversationWorkerOptions {
     userId: string,
   ) => Promise<SlackRequesterProfile | null | undefined>;
   resumeAwaitingContinuation: (conversationId: string) => Promise<boolean>;
-  metadataStore?: ConversationMetadataStore;
+  conversationStore?: ConversationStore;
   runtime: Pick<
     SlackTurnRuntime<unknown>,
     "handleNewMention" | "handleSubscribedMessage"
@@ -213,7 +213,7 @@ export function createSlackConversationWorker(
     const records = getPendingRecords(
       await getConversationWorkState({
         conversationId: context.conversationId,
-        metadataStore: options.metadataStore,
+        conversationStore: options.conversationStore,
         state,
       }),
     );
@@ -284,7 +284,7 @@ export function createSlackConversationWorker(
             conversationId: context.conversationId,
             inboundMessageIds: initialInboundMessageIds,
             leaseToken: context.leaseToken,
-            metadataStore: options.metadataStore,
+            conversationStore: options.conversationStore,
             state,
           });
           initialMessagesPersisted = marked;
