+++ packages/junior/src/handlers/agent-dispatch.ts
@@ -1,12 +1,12 @@
 import { logException } from "@/chat/logging";
 import { runAgentDispatchSlice } from "@/chat/agent-dispatch/runner";
 import { verifyDispatchCallbackRequest } from "@/chat/agent-dispatch/signing";
-import type { ConversationMetadataStore } from "@/chat/metadata/store";
+import type { ConversationStore } from "@/chat/conversations/store";
 import type { SandboxEgressTracePropagationConfig } from "@/chat/sandbox/egress-tracing";
 import type { WaitUntilFn } from "@/handlers/types";
 
 interface AgentDispatchHandlerOptions {
-  metadataStore?: ConversationMetadataStore;
+  conversationStore?: ConversationStore;
   tracePropagation?: SandboxEgressTracePropagationConfig;
 }
 
@@ -23,7 +23,7 @@ export async function POST(
 
   waitUntil(() =>
     runAgentDispatchSlice(payload, {
-      metadataStore: options.metadataStore,
+      conversationStore: options.conversationStore,
       tracePropagation: options.tracePropagation,
     }).catch((error) => {
       logException(
